chore: extract shared localStorage log utility from transactionLog and adminLog
Extract duplicated canUseStorage/readAllEntries/writeAllEntries logic
into a new src/lib/localLog.ts utility. Both transactionLog.ts and
adminLog.ts now import from the shared module.

Refs #838
